requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=pypi_0
apache-beam=2.30.0=pypi_0
astunparse=1.6.3=pypi_0
attrs=21.2.0=pypi_0
avro-python3=1.9.2.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py36ha15d459_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
contextlib2=21.6.0=pypi_0
crcmod=1.7=pypi_0
cudatoolkit=11.2.2=h933977f_8
cudnn=8.2.1.32=h754d62a_0
cycler=0.10.0=pypi_0
cython=0.29.23=pypi_0
dataclasses=0.8=pypi_0
decorator=5.0.9=pyhd3eb1b0_0
dill=0.3.1.1=pypi_0
dm-tree=0.1.6=pypi_0
docopt=0.6.2=pypi_0
fastavro=1.4.2=pypi_0
flatbuffers=1.12=pypi_0
future=0.18.2=pypi_0
gast=0.4.0=pypi_0
gin-config=0.4.0=pypi_0
git=2.30.2=h57928b3_1
google-api-core=1.30.0=pypi_0
google-api-python-client=2.11.0=pypi_0
google-auth=1.32.1=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-cloud-bigquery=2.20.0=pypi_0
google-cloud-core=1.7.1=pypi_0
google-crc32c=1.1.2=pypi_0
google-pasta=0.2.0=pypi_0
google-resumable-media=1.3.1=pypi_0
googleapis-common-protos=1.53.0=pypi_0
grpcio=1.34.1=pypi_0
h5py=3.1.0=pypi_0
hdfs=2.6.0=pypi_0
httplib2=0.19.1=pypi_0
idna=2.10=pypi_0
importlib-metadata=4.6.0=pypi_0
importlib-resources=5.2.0=pypi_0
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py36_0
joblib=1.0.1=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py36haa95532_0
kaggle=1.5.12=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
libprotobuf=3.17.2=h7755175_0
libsodium=1.0.18=h62dcd97_0
lvis=0.5.3=pypi_0
lxml=4.6.3=pypi_0
markdown=3.3.4=pypi_0
matplotlib=3.3.4=pypi_0
numpy=1.19.5=pypi_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.1=pypi_0
object-detection=0.1=pypi_0
opencv-python=4.5.2.54=pypi_0
opencv-python-headless=4.5.2.54=pypi_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pypi_0
pandas=1.1.5=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.0=pypi_0
pip=21.1.3=py36haa95532_0
portalocker=2.0.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.17=pyh06a4308_0
proto-plus=1.19.0=pypi_0
protobuf=3.17.2=py36he2d232f_0
psutil=5.8.0=pypi_0
py-cpuinfo=8.0.0=pypi_0
pyarrow=3.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybind11=2.6.2=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=pypi_0
pydot=1.4.2=pypi_0
pygments=2.9.0=pyhd3eb1b0_0
pymongo=3.11.4=pypi_0
pyparsing=2.4.7=pypi_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-slugify=5.0.2=pypi_0
python_abi=3.6=2_cp36m
pytz=2021.1=pypi_0
pywin32=301=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py36hd77b12b_1
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
sacrebleu=1.5.1=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
sentencepiece=0.1.96=pypi_0
seqeval=1.2.2=pypi_0
setuptools=52.0.0=py36haa95532_0
six=1.15.0=pypi_0
soupsieve=2.0.1=py_1
sqlite=3.36.0=h2bbff1b_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-addons=0.13.0=pypi_0
tensorflow-datasets=4.3.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-metadata=1.1.0=pypi_0
tensorflow-model-optimization=0.6.0=pypi_0
termcolor=1.1.0=pypi_0
text-unidecode=1.3=pypi_0
tf-models-official=2.5.0=pypi_0
tf-slim=1.1.0=pypi_0
tflite-support-nightly=0.1.0.dev20210706=pypi_0
threadpoolctl=2.1.0=pypi_0
tornado=6.1=py36h2bbff1b_0
tqdm=4.61.1=pypi_0
traitlets=4.3.3=py36_0
typeguard=2.12.1=pypi_0
typing-extensions=3.7.4.3=pypi_0
uritemplate=3.0.1=pypi_0
urllib3=1.26.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.1=pypi_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pypi_0
zlib=1.2.11=h62dcd97_1010
